requirements.txt
# Data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # For Excel file support

# SQL query execution
pandasql>=0.7.3

# Google Gemini AI
google-genai>=0.21.0

# Web interface 
streamlit>=1.28.0

python-dotenv>=1.0.0  # For managing API keys
plotly>=5.17.0  # For visualizations
matplotlib>=3.7.0  # For plots
seaborn>=0.12.0  # For statistical visualizations
torch